+++ docs/examples/authorization_code_grant.py
@@ -68,12 +68,12 @@ class TestSiteAdapter(AuthorizationCodeGrantSiteAdapter):
 </html>
     """
 
-    def render_auth_page(self, request, response, environ, scopes):
+    def render_auth_page(self, request, response, environ, scopes, client):
         response.body = self.CONFIRMATION_TEMPLATE
 
         return response
 
-    def authenticate(self, request, environ, scopes):
+    def authenticate(self, request, environ, scopes, client):
         if request.method == "POST":
             if request.post_param("confirm") == "1":
                 return